+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Support for `Accept`/`Content-type` headers and for the content-type aware combinators in *servant-0.3*
 * Export `toApplication` from `Servant.Server` (https://github.com/haskell-servant/servant-server/pull/29)
 * Support other Monads than just `EitherT (Int, String) IO` (https://github.com/haskell-servant/servant-server/pull/21)
+* Make methods return status code 204 if they return () (https://github.com/haskell-servant/servant-server/issues/28)
 
 0.2.4
 -----